Add files via upload
This python script will answer the Practice problem at the end of chapter 8.  This was done by a relative newcomer to Python so it may be able to be optimized, but it will do the job.  

The code used SHOULD work on either Windows or Mac but only for Python 3+